requirements.txt
Pillow>=9.5.0
aiohttp
piexif
watchdog
asyncio